% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agilaire_flags}
\alias{agilaire_flags}
\title{agilaire_flags dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 49 rows and 4 columns.
}
\usage{
agilaire_flags
}
\description{
agilaire_flags dataset
}
\author{
Simeon \email{simeon.ng@alaska.gov}
}
\keyword{data}
