% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agileweb.R
\name{agileweb_dt}
\alias{agileweb_dt}
\title{agileweb_dt}
\usage{
agileweb_dt(datetime)
}
\arguments{
\item{datetime}{Datetime in POSIXct format.}
}
\value{
Datetime in required Agileweb API format.
}
\description{
agileweb_dt
}
\examples{
end <- agileweb_dt(lubridate::now())
}
