% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{avg_pct_diff}
\alias{avg_pct_diff}
\title{Calculate average percent difference between two sensors}
\usage{
avg_pct_diff(data, col1, col2)
}
\arguments{
\item{data}{Dataset with wide pivoted sensor/pollutant columns}

\item{col1}{Name of column of sample measurements of sensor 1}

\item{col2}{Name of column of sample measurements of sensor 2}
}
\value{
Average percent difference of two sensors
}
\description{
Calculate average percent difference between two sensors
}
