% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_completeness}
\alias{check_completeness}
\title{Check Completeness of Dataset}
\usage{
check_completeness(data, pollutant)
}
\arguments{
\item{data}{AQ dataset}

\item{pollutant}{Pollutant to analyze completeness}
}
\value{
Boolean of if data meets completeness
}
\description{
Check Completeness of Dataset
}
