% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_co_qt_bam}
\alias{clean_co_qt_bam}
\title{Cleans a Joined Quant/BAM dataset for CO}
\usage{
clean_co_qt_bam(data)
}
\arguments{
\item{data}{A joined Quant/BAM dataset}
}
\value{
A cleaned CO dataset
}
\description{
Cleans a Joined Quant/BAM dataset for CO
}
