% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_garden_quant}
\alias{clean_garden_quant}
\title{Cleans Garden Quant Imported from AA}
\usage{
clean_garden_quant(data, pod)
}
\arguments{
\item{data}{Tibble imported using the read_quant_flagged() function}

\item{pod}{MODULAIR Pod #}
}
\value{
Clean tibble of Quant data.
}
\description{
Cleans Garden Quant Imported from AA
}
