% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_pm10}
\alias{clean_pm10}
\title{Extracts PM10 data from an Agileweb query.}
\usage{
clean_pm10(data)
}
\arguments{
\item{data}{Tibble from an agileweb query.}
}
\value{
Clean tibble with date and PM10 sample.
}
\description{
Extracts PM10 data from an Agileweb query.
}
