% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_pm25}
\alias{clean_pm25}
\title{Extracts PM2.5 data from an Agileweb query.}
\usage{
clean_pm25(data)
}
\arguments{
\item{data}{Tibble from an agileweb query.}
}
\value{
Clean tibble with date and PM2.5 sample.
}
\description{
Extracts PM2.5 data from an Agileweb query.
}
