% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_pm25_qt_bam}
\alias{clean_pm25_qt_bam}
\title{Cleans a Joined Quant/BAM dataset for PM2.5}
\usage{
clean_pm25_qt_bam(data)
}
\arguments{
\item{data}{A joined Quant/BAM dataset}
}
\value{
A cleaned PM2.5 dataset
}
\description{
Cleans a Joined Quant/BAM dataset for PM2.5
}
