% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_qt_to_qt}
\alias{clean_qt_to_qt}
\title{Create a Joined Quant Dataset}
\usage{
clean_qt_to_qt(quant1, quant2, suffixes)
}
\arguments{
\item{quant1}{Quant 1 dataset}

\item{quant2}{Quant 2 dataset}

\item{suffixes}{Quant SNs}
}
\value{
A cleaned Quant/Quant dataset
}
\description{
Create a Joined Quant Dataset
}
