% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_regulatory}
\alias{clean_regulatory}
\title{Cleans Regulatory Data Imported from AA}
\usage{
clean_regulatory(data)
}
\arguments{
\item{data}{Tibble imported using the read_regulatory_flagged() function}
}
\value{
Clean tibble of regulatory data.
}
\description{
Cleans Regulatory Data Imported from AA
}
