% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_relhum}
\alias{clean_relhum}
\title{Extracts AMBTEMP data from an Agileweb query.}
\usage{
clean_relhum(data)
}
\arguments{
\item{data}{Tibble from an agileweb query.}
}
\value{
Clean tibble with date and AMBTEMP sample.
}
\description{
Extracts AMBTEMP data from an Agileweb query.
}
