% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{convert_aqi_pm25}
\alias{convert_aqi_pm25}
\title{Converts a value to an AQI value}
\usage{
convert_aqi_pm25(value)
}
\arguments{
\item{value}{NowCast calculated value}
}
\value{
An AQI value
}
\description{
Converts a value to an AQI value
}
