% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lm_caption}
\alias{lm_caption}
\title{Caption for lm plot}
\usage{
lm_caption(slope, int, r2)
}
\arguments{
\item{slope}{slope value}

\item{int}{y-intercept value}

\item{r2}{r-squared value}
}
\value{
linear equation caption string
}
\description{
Caption for lm plot
}
