% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncore_quants}
\alias{ncore_quants}
\title{ncore_quants dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4875 rows and 57 columns.
}
\usage{
ncore_quants
}
\description{
ncore_quants dataset
}
\author{
Simeokn \email{simeon.ng@alaska.gov}
}
\keyword{data}
