% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_co_ts_monitor}
\alias{plot_co_ts_monitor}
\title{CO Time Series plot for any Monitor(s)}
\usage{
plot_co_ts_monitor(data, title)
}
\arguments{
\item{data}{CO Dataset. Must be in 'Date | CO | Monitor' form.}

\item{title}{Title of plot.}
}
\value{
A CO time series ggplot
}
\description{
CO Time Series plot for any Monitor(s)
}
