% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_diurnal}
\alias{plot_diurnal}
\title{PM2.5 Diurnal Plot}
\usage{
plot_diurnal(data, pollutant, sites, title, statistic)
}
\arguments{
\item{data}{Data for plot.}

\item{pollutant}{Pollutant name as a string (usually column name).}

\item{sites}{Site name or list of sites.}

\item{title}{A string for the title of the plot.}

\item{statistic}{A string "median" or "mean"}
}
\value{
A Compilation of Diurnal Plots.
}
\description{
PM2.5 Diurnal Plot
}
\examples{
\dontrun{
plot_diurnal(quant_pm25, "pm25", c("Bethel", "Napaskiak"), "Example Plot", "median")
}

}
