% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lm}
\alias{plot_lm}
\title{Linear Model Plot for any Pollutant & Monitor(s)}
\usage{
plot_lm(data, title, xlab, ylab, caption)
}
\arguments{
\item{data}{Pollutant Dataset. Must be in 'Date | Quant | BAM' form.}

\item{title}{Title of the plot.}

\item{xlab}{X axis label.}

\item{ylab}{Y axis label.}

\item{caption}{Equation caption.}
}
\value{
A Linear Model ggplot
}
\description{
Linear Model Plot for any Pollutant & Monitor(s)
}
