% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lm_quant}
\alias{plot_lm_quant}
\title{Linear Model Plot for any Pollutant & Monitor(s)}
\usage{
plot_lm_quant(data, title, ind, dep, pollutant)
}
\arguments{
\item{data}{Pollutant Dataset. Must be in 'Date | Monitor | Pollutant' form.}

\item{title}{Title of the plot.}

\item{ind}{Independent variable.}

\item{dep}{Dependent variable.}

\item{pollutant}{Pollutant to compare.}
}
\value{
A Linear Model ggplot
}
\description{
Linear Model Plot for any Pollutant & Monitor(s)
}
