% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_loess}
\alias{plot_loess}
\title{Loess Regression Plot for any Pollutant & Monitor(s)}
\usage{
plot_loess(data, x, y, title, xlab, ylab, caption = NULL)
}
\arguments{
\item{data}{Pollutant Dataset. Must be in 'Date | Quant | BAM' form.}

\item{x}{Independent variable}

\item{y}{Dependent variable}

\item{title}{Title of the plot.}

\item{xlab}{X axis label.}

\item{ylab}{Y axis label.}

\item{caption}{Equation caption.}
}
\value{
A Loess Smooth Line ggplot
}
\description{
Loess Regression Plot for any Pollutant & Monitor(s)
}
