% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_no_ts_monitor}
\alias{plot_no_ts_monitor}
\title{NO Time Series plot for any Monitor(s)}
\usage{
plot_no_ts_monitor(data, title)
}
\arguments{
\item{data}{NO Dataset. Must be in 'Date | NO | Monitor' form.}

\item{title}{Title of plot.}
}
\value{
A NO time series ggplot
}
\description{
NO Time Series plot for any Monitor(s)
}
