% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_o3_ts_monitor}
\alias{plot_o3_ts_monitor}
\title{O3 Time Series plot for any Monitor(s)}
\usage{
plot_o3_ts_monitor(data, title)
}
\arguments{
\item{data}{O3 Dataset. Must be in 'Date | O3 | Monitor' form.}

\item{title}{Title of plot.}
}
\value{
A O3 time series ggplot
}
\description{
O3 Time Series plot for any Monitor(s)
}
