% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pm10_ts_monitor}
\alias{plot_pm10_ts_monitor}
\title{PM10 Time Series plot for any Monitor(s)}
\usage{
plot_pm10_ts_monitor(data, title)
}
\arguments{
\item{data}{PM10 Dataset. Must be in 'Date | PM10 | Monitor' form.}

\item{title}{Title of plot.}
}
\value{
A PM10 time series ggplot
}
\description{
PM10 Time Series plot for any Monitor(s)
}
