% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pm25_ts_monitor}
\alias{plot_pm25_ts_monitor}
\title{PM2.5 Time Series Plot for any Monitor(s)}
\usage{
plot_pm25_ts_monitor(data, title)
}
\arguments{
\item{data}{PM2.5 Dataset. Must be in 'Date | PM25 | Monitor' form.}

\item{title}{Title of the plot.}
}
\value{
A PM2.5 time series ggplot
}
\description{
PM2.5 Time Series Plot for any Monitor(s)
}
