% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_quant_pm25}
\alias{plot_quant_pm25}
\title{Quant PM25 Temporal Plot}
\usage{
plot_quant_pm25(data, town, cols = NULL, ...)
}
\arguments{
\item{data}{Quant data in Agilaire basic data export format.}

\item{town}{Quant town location.}

\item{cols}{Optional plot color.}
}
\value{
A PM2.5 temporal plot.
}
\description{
Quant PM25 Temporal Plot
}
\examples{
data <- read_aa("C:/Data Analysis/Testing/data/haines_quant.csv")
plot_quant_pm25(data, "Haines")
}
