% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rq}
\alias{plot_rq}
\title{Median Regression Plot for any Pollutant & Monitor(s)}
\usage{
plot_rq(data, x, y, title, xlab, ylab, caption = NULL)
}
\arguments{
\item{data}{Pollutant Dataset. Must be in 'Date | Quant | BAM' form.}

\item{title}{Title of the plot.}

\item{xlab}{X axis label.}

\item{ylab}{Y axis label.}

\item{caption}{Equation caption.}
}
\value{
A Median Regression ggplot
}
\description{
Median Regression Plot for any Pollutant & Monitor(s)
}
