% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_wind_rose}
\alias{plot_wind_rose}
\title{Wind Rose plot to DEC AMQA Standard}
\usage{
plot_wind_rose(data, title)
}
\arguments{
\item{data}{Wind speed and wind direction data}

\item{title}{Title for plot}
}
\value{
An openair Windrose
}
\description{
Wind Rose plot to DEC AMQA Standard
}
