% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quant_460}
\alias{quant_460}
\title{quant_460 agilaire query quant dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5201 rows and 3 columns.
}
\usage{
quant_460
}
\description{
quant_460 agilaire query quant dataset
}
\author{
Simeon \email{simeon.ng@alaska.gov}
}
\keyword{data}
