% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aa.R
\name{read_aa}
\alias{read_aa}
\title{Read an Agilaire File}
\usage{
read_aa(file)
}
\arguments{
\item{file}{The path to the file.}
}
\value{
Agilaire data in tibble form.
}
\description{
Reads a basic data export file from Agilaire/AirVision and reformats
data into useable form.
}
\examples{
file_path <- "C:/Data Analysis/Testing/data/BasicDataExportReport.csv"
basic_data_aa <- read_aa(file_path)
gt::gt_preview(basic_data_aa, top_n = 10, incl_rownums = FALSE)
}
