% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_quant_flagged}
\alias{read_quant_flagged}
\title{Reads an Agilaire Quant Flagged Data Export}
\usage{
read_quant_flagged(file)
}
\arguments{
\item{file}{The path to the file.}
}
\value{
Agilaire data in tibble form.
}
\description{
Reads an Agilaire Quant Flagged Data Export
}
