% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_regulatory_flagged}
\alias{read_regulatory_flagged}
\title{Reads an Agilaire Regulatory Flagged Data Export}
\usage{
read_regulatory_flagged(file, site)
}
\arguments{
\item{file}{The path to the file.}

\item{site}{The Regulatory site name as a string. This should match the site name exported from AA.}
}
\value{
Agilaire data in tibble form.
}
\description{
Reads an Agilaire Regulatory Flagged Data Export
}
