% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{valid_nowcast}
\alias{valid_nowcast}
\title{Valid NowCast hour if 2 out of the last 3 hours of data are valid.}
\usage{
valid_nowcast(data)
}
\arguments{
\item{data}{PM Dataset}
}
\value{
Boolean if hour is valid
}
\description{
Valid NowCast hour if 2 out of the last 3 hours of data are valid.
}
