% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear_regression}
\alias{linear_regression}
\title{Linear Regression Analysis}
\usage{
linear_regression(data, x, y)
}
\arguments{
\item{data}{Data to analyze}

\item{x}{String of column of independent variable}

\item{y}{String of column of dependent variable}
}
\value{
Vector of lm coefficients
}
\description{
Linear Regression Analysis
}
