% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_diurnal}
\alias{plot_diurnal}
\title{Diurnal Plot}
\usage{
plot_diurnal(data, pollutant, sites, title, statistic)
}
\arguments{
\item{data}{Data for plot.}

\item{pollutant}{Pollutant name as a string.}

\item{sites}{Site name or list of sites.}

\item{title}{Title for plot.}

\item{statistic}{Select mean or median.}
}
\value{
A Compilation of Diurnal Plots.
}
\description{
Diurnal Plot
}
