% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_diurnal_cc}
\alias{plot_diurnal_cc}
\title{Diurnal Plot - Community Call}
\usage{
plot_diurnal_cc(data, site, start, end, statistic)
}
\arguments{
\item{data}{Quant data for plot}

\item{site}{Quant sensor name}

\item{start}{Start date}

\item{end}{End date}

\item{statistic}{Median or Mean}
}
\value{
Diurnal ggplot
}
\description{
Diurnal Plot - Community Call
}
