% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_wind_rose}
\alias{plot_wind_rose}
\title{DEC AMQA Wind Rose}
\usage{
plot_wind_rose(data, title)
}
\arguments{
\item{data}{Wind speed and wind direction data. Must have column names ("date", "wd", "ws").}

\item{title}{Title for the figure.}
}
\value{
An openair Windrose
}
\description{
DEC AMQA Wind Rose
}
